+++ src/Tracking/Consent/AllowAllTrackingConsentChecker.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Zhortein\SeoTrackingBundle\Tracking\Consent;
+
+use Symfony\Component\HttpFoundation\Request;
+
+final readonly class AllowAllTrackingConsentChecker implements TrackingConsentCheckerInterface
+{
+    public function isGranted(?Request $request): bool
+    {
+        return true;
+    }
+}